Refine playback transport module boundaries
