+++ src/app/controller/playback/transport/looping.rs
@@ -1,4 +1,13 @@
 use super::*;
+
+/// Persistence side effects for loop toggle metadata.
+mod persistence;
+use persistence::persist_loop_toggle_markers;
+
+/// Loop transport policy tests.
+#[cfg(test)]
+mod tests;
+
 pub(crate) fn toggle_loop(controller: &mut AppController) {
     if controller.ui.waveform.loop_lock_enabled {
         controller.set_loop_lock_enabled(false);
@@ -73,167 +82,6 @@ fn cycle_loop_lock_state(controller: &mut AppController) -> LoopToggleState {
     }
 }
 
-/// Browser action rows used for multi-sample loop/BPM metadata writes.
-struct LoopActionTargets {
-    /// Primary browser row (loaded sample when visible).
-    primary_row: Option<usize>,
-    /// Action target paths (selection plus primary when needed).
-    paths: Vec<std::path::PathBuf>,
-}
-
-/// Resolve action paths targeted by loop metadata updates.
-fn loop_action_targets(controller: &mut AppController) -> LoopActionTargets {
-    let loaded_path = controller
-        .sample_view
-        .wav
-        .loaded_audio
-        .as_ref()
-        .map(|audio| audio.relative_path.clone());
-    let primary_row = loaded_path
-        .as_ref()
-        .and_then(|path| controller.visible_row_for_path(path));
-    let paths = primary_row
-        .map(|row| controller.browser_action_paths_from_primary(row))
-        .unwrap_or_default();
-    LoopActionTargets { primary_row, paths }
-}
-
-/// Persist loop marker state to selected browser rows or loaded-sample fallback.
-fn persist_loop_toggle_markers(controller: &mut AppController, state: LoopToggleState) {
-    let action_targets = loop_action_targets(controller);
-    if !action_targets.paths.is_empty() {
-        persist_browser_loop_markers(controller, &action_targets, state);
-    } else {
-        persist_loaded_sample_loop_marker(controller, state.loop_enabled);
-        if state.toggled_to_enabled() {
-            persist_loaded_sample_bpm(controller);
-        }
-    }
-}
-
-/// Persist loop markers (and initial BPM when enabling) across targeted browser rows.
-fn persist_browser_loop_markers(
-    controller: &mut AppController,
-    action_targets: &LoopActionTargets,
-    state: LoopToggleState,
-) {
-    let primary_row = action_targets.primary_row.unwrap_or(0);
-    if let Err(err) = controller.set_loop_marker_browser_sample_paths(
-        &action_targets.paths,
-        state.loop_enabled,
-        primary_row,
-    ) {
-        tracing::warn!("Failed to update loop markers for browser samples: {err}");
-    }
-    if state.toggled_to_enabled()
-        && let Some(bpm) = controller.ui.waveform.bpm_value
-        && bpm.is_finite()
-        && bpm > 0.0
-    {
-        if let Some(source_id) = controller.selected_source_id() {
-            let cache = controller
-                .ui_cache
-                .browser
-                .bpm_values
-                .entry(source_id)
-                .or_default();
-            for path in &action_targets.paths {
-                cache.insert(path.clone(), Some(bpm));
-            }
-            controller.mark_browser_row_metadata_projection_revision_dirty();
-        }
-        if let Err(err) =
-            controller.set_bpm_browser_sample_paths(&action_targets.paths, bpm, primary_row)
-        {
-            tracing::warn!("Failed to save BPM to browser samples: {err}");
-        }
-    }
-}
-
-/// Persist loop marker state for the loaded sample when no browser rows are actionable.
-fn persist_loaded_sample_loop_marker(controller: &mut AppController, loop_enabled: bool) {
-    let loop_marker_update =
-        controller
-            .sample_view
-            .wav
-            .loaded_audio
-            .as_ref()
-            .and_then(|loaded_audio| {
-                controller
-                    .library
-                    .sources
-                    .iter()
-                    .find(|source| source.id == loaded_audio.source_id)
-                    .map(|source| (source.clone(), loaded_audio.relative_path.clone()))
-            });
-    if let Some((source, relative_path)) = loop_marker_update
-        && let Err(err) =
-            controller.set_sample_looped_for_source(&source, &relative_path, loop_enabled, false)
-    {
-        tracing::warn!("Failed to update loop marker: {err}");
-    }
-}
-
-/// Persist BPM metadata for the loaded sample when loop-enable falls back to sample scope.
-fn persist_loaded_sample_bpm(controller: &mut AppController) {
-    let Some(bpm) = controller
-        .ui
-        .waveform
-        .bpm_value
-        .filter(|bpm| bpm.is_finite() && *bpm > 0.0)
-    else {
-        return;
-    };
-    let bpm_update = controller
-        .sample_view
-        .wav
-        .loaded_audio
-        .as_ref()
-        .and_then(|loaded_audio| {
-            controller
-                .library
-                .sources
-                .iter()
-                .find(|source| source.id == loaded_audio.source_id)
-                .map(|source| (source.clone(), loaded_audio.relative_path.clone()))
-        });
-    let Some((source, relative_path)) = bpm_update else {
-        return;
-    };
-    let before_bpm = controller
-        .ui_cache
-        .browser
-        .bpm_values
-        .get(&source.id)
-        .and_then(|cache| cache.get(&relative_path).copied().flatten());
-    controller
-        .ui_cache
-        .browser
-        .bpm_values
-        .entry(source.id.clone())
-        .or_default()
-        .insert(relative_path.clone(), Some(bpm));
-    controller.queue_metadata_mutation(
-        &source,
-        Vec::new(),
-        vec![
-            crate::app::controller::jobs::AnalysisMetadataMutationOp::SetBpm {
-                relative_path: relative_path.clone(),
-                bpm: Some(bpm),
-            },
-        ],
-        vec![
-            crate::app::controller::state::runtime::MetadataRollback::Bpm {
-                relative_path,
-                before_bpm,
-                expected_bpm: Some(bpm),
-            },
-        ],
-        false,
-    );
-    controller.mark_browser_row_metadata_projection_revision_dirty();
-}
-
 /// Apply loop-toggle playback policy (restart looping playback or defer disable).
 fn apply_loop_toggle_playback_policy(controller: &mut AppController, state: LoopToggleState) {
     match loop_playback_policy(state) {
@@ -298,159 +146,4 @@ fn loop_restart_start_override(controller: &AppController, progress: Option<f32>
 /// Return true when a valid playback selection should drive loop restart position.
 fn has_loop_playback_selection(controller: &AppController) -> bool {
     super::playback_audition_selection(controller).is_some()
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::app::controller::test_support;
-    use std::path::Path;
-    use std::time::{Duration, Instant};
-
-    #[test]
-    fn toggle_loop_enable_clears_pending_disable_deadline() {
-        let (mut controller, _source) = test_support::dummy_controller();
-        controller.audio.pending_loop_disable_at = Some(Instant::now() + Duration::from_secs(1));
-        controller.ui.waveform.loop_enabled = false;
-
-        toggle_loop(&mut controller);
-
-        assert!(controller.ui.waveform.loop_enabled);
-        assert!(controller.audio.pending_loop_disable_at.is_none());
-    }
-
-    #[test]
-    fn loop_playback_policy_maps_toggle_transitions() {
-        assert_eq!(
-            loop_playback_policy(LoopToggleState {
-                was_looping: false,
-                loop_enabled: true,
-            }),
-            LoopPlaybackPolicy::RestartIfPlaying
-        );
-        assert_eq!(
-            loop_playback_policy(LoopToggleState {
-                was_looping: true,
-                loop_enabled: false,
-            }),
-            LoopPlaybackPolicy::DeferDisableAfterCycle
-        );
-        assert_eq!(
-            loop_playback_policy(LoopToggleState {
-                was_looping: false,
-                loop_enabled: false,
-            }),
-            LoopPlaybackPolicy::None
-        );
-    }
-
-    #[test]
-    fn toggle_loop_unlocks_before_persisting_normal_toggle() {
-        let (mut controller, _source) = test_support::dummy_controller();
-        controller.ui.waveform.loop_enabled = true;
-        controller.set_loop_lock_enabled(true);
-
-        toggle_loop(&mut controller);
-
-        assert!(!controller.ui.waveform.loop_lock_enabled);
-        assert!(!controller.ui.waveform.loop_enabled);
-    }
-
-    #[test]
-    fn toggle_loop_lock_enters_locked_on_from_unlocked() {
-        let (mut controller, _source) = test_support::dummy_controller();
-        controller.ui.waveform.loop_enabled = false;
-        controller.set_loop_lock_enabled(false);
-
-        toggle_loop_lock(&mut controller);
-
-        assert!(controller.ui.waveform.loop_lock_enabled);
-        assert!(controller.ui.waveform.loop_enabled);
-    }
-
-    #[test]
-    fn toggle_loop_lock_cycles_locked_on_and_off() {
-        let (mut controller, _source) = test_support::dummy_controller();
-        controller.ui.waveform.loop_enabled = false;
-
-        toggle_loop_lock(&mut controller);
-        assert!(controller.ui.waveform.loop_lock_enabled);
-        assert!(controller.ui.waveform.loop_enabled);
-
-        toggle_loop_lock(&mut controller);
-        assert!(controller.ui.waveform.loop_lock_enabled);
-        assert!(!controller.ui.waveform.loop_enabled);
-
-        toggle_loop_lock(&mut controller);
-        assert!(controller.ui.waveform.loop_lock_enabled);
-        assert!(controller.ui.waveform.loop_enabled);
-    }
-
-    #[test]
-    fn toggle_loop_lock_does_not_persist_sample_loop_marker() {
-        let (mut controller, source) =
-            test_support::prepare_with_source_and_wav_entries(vec![test_support::sample_entry(
-                "locked_override.wav",
-                crate::sample_sources::Rating::NEUTRAL,
-            )]);
-        let db = controller.cache_db(&source).expect("db");
-        db.set_looped(Path::new("locked_override.wav"), false)
-            .expect("seed loop marker");
-        controller.sample_view.wav.loaded_audio =
-            Some(crate::app::controller::state::audio::LoadedAudio {
-                source_id: source.id.clone(),
-                root: source.root.clone(),
-                relative_path: Path::new("locked_override.wav").to_path_buf(),
-                bytes: Vec::new().into(),
-                duration_seconds: 1.0,
-                sample_rate: 48_000,
-            });
-
-        toggle_loop_lock(&mut controller);
-
-        assert_eq!(
-            db.looped_for_path(Path::new("locked_override.wav"))
-                .expect("load loop marker"),
-            Some(false)
-        );
-    }
-
-    #[test]
-    fn loop_restart_start_override_uses_selection_span_when_available() {
-        let (mut controller, _source) = test_support::dummy_controller();
-        controller.ui.waveform.selection = Some(SelectionRange::new(0.2, 0.6));
-
-        let start_override = loop_restart_start_override(&controller, Some(0.4));
-
-        assert_eq!(start_override, None);
-    }
-
-    #[test]
-    fn loop_restart_start_override_priority_chain_is_stable() {
-        let (mut controller, _source) = test_support::dummy_controller();
-
-        let with_progress = loop_restart_start_override(&controller, Some(0.42));
-        assert_eq!(with_progress, Some(f64::from(0.42_f32)));
-
-        controller.ui.waveform.playhead.visible = true;
-        controller.ui.waveform.playhead.position = 0.37;
-        let with_playhead = loop_restart_start_override(&controller, None);
-        assert_eq!(
-            with_playhead,
-            Some(f64::from(controller.ui.waveform.playhead.position))
-        );
-
-        controller.ui.waveform.playhead.visible = false;
-        controller.ui.waveform.cursor = Some(0.55);
-        controller.ui.waveform.last_start_marker = Some(0.19);
-        let with_cursor = loop_restart_start_override(&controller, None);
-        assert_eq!(with_cursor, controller.ui.waveform.cursor.map(f64::from));
-
-        controller.ui.waveform.cursor = None;
-        let with_last_marker = loop_restart_start_override(&controller, None);
-        assert_eq!(
-            with_last_marker,
-            controller.ui.waveform.last_start_marker.map(f64::from)
-        );
-    }
 }
